refactor(core): models package name changed with game
