+++ core/src/main/java/org/catanuniverse/core/game/Configuration.java
@@ -3,7 +3,7 @@
 	22015094 - Idil Saglam
 	 - Abderrahim Arous
 */
-package org.catanuniverse.core.models;
+package org.catanuniverse.core.game;
 
 public class Configuration {
