+++ core/src/main/java/org/catanuniverse/core/game/Player.java
@@ -3,6 +3,6 @@
 	22015094 - Idil Saglam
 	 - Abderrahim Arous
 */
-package org.catanuniverse.core.models;
+package org.catanuniverse.core.game;
 
 public class Player {}
