+++ core/src/test/java/org/catanuniverse/core/game/PositionsTest.java
@@ -3,7 +3,7 @@
 	22015094 - Idil Saglam
 	 - Abderrahim Arous
 */
-package org.catanuniverse.core.models;
+package org.catanuniverse.core.game;
 
 import static org.junit.jupiter.params.provider.Arguments.arguments;
